feat: Refactor dashboard to include new pages
