+++ app/src/components/dash.rs
@@ -6,11 +6,11 @@ use crate::utils::structs::UpdateUser;
 
 /// Renders the home page of your application.
 #[component(transparent)]
-pub fn Panel() -> impl IntoView {
+pub fn Dash() -> impl IntoView {
     use crate::components::admin::*;
-    use crate::components::panel_content::*;
-    use crate::components::panel_nav::*;
-    use crate::components::projects::*;
+    use crate::components::dash_landing_content::*;
+    use crate::components::dash_nav::*;
+    use crate::components::search::*;
     use crate::components::tokens::*;
 
     let update_user = use_context::<UpdateUser>().expect("user_state not set");
@@ -19,20 +19,19 @@ pub fn Panel() -> impl IntoView {
 
     view! {
         <Route
-            path="panel"
+            path="dash"
             view=move || {
                 view! {
-                    <PanelNav/>
+                    <DashNav/>
                     <Outlet/>
                 }
             }
         >
-
-            <Route path="projects" view=move || view! { <ProjectsOverview/> }/>
-            <Route path="collection" view=move || "Collections".into_view()/>
+            <Route path="search" view=move || view! { <Search/> }/>
+            <Route path="resources" view=move || "Resources".into_view()/>
             <Route path="admin" view=move || view! { <AdminOverview/> }/>
             <Route path="tokens" view=move || view! { <TokensOverview/> }/>
-            <Route path="" view=move || view! { <PanelContent/> }/>
+            <Route path="" view=move || view! { <DashLanding/> }/>
         </Route>
     }
 }
