+++ app/src/components/mod.rs
@@ -6,14 +6,14 @@ pub mod alert_modal;
 pub mod create_object;
 pub mod create_project;
 pub mod create_token;
+pub mod dash;
+pub mod dash_landing_content;
+pub mod dash_nav;
 pub mod footer;
 pub mod header;
 pub mod imprint;
 pub mod main_body;
 pub mod objects;
-pub mod panel;
-pub mod panel_content;
-pub mod panel_nav;
 pub mod project;
 pub mod project_admin;
 pub mod projects;
